move `ZoomSpeed` and `CameraSpeeds` to module `robot_cameraman.camera_speeds`
