+++ robot_cameraman/tracking.py
@@ -1,14 +1,14 @@
 import logging
 import time
 from abc import abstractmethod
-from dataclasses import dataclass
-from enum import Enum, auto, IntEnum
+from enum import Enum, auto
 from logging import Logger
 from typing import Optional
 
 from typing_extensions import Protocol
 
 from robot_cameraman.box import Box, TwoPointsBox, Point
+from robot_cameraman.camera_speeds import ZoomSpeed, CameraSpeeds
 from robot_cameraman.live_view import ImageSize
 
 logger: Logger = logging.getLogger(__name__)
@@ -42,42 +42,6 @@ class Destination:
         self.min_size_box.x = x - self.min_size_box.width / 2
         self.min_size_box.y = y - self.min_size_box.height / 2
         self.min_size_box.center.set(x, y)
-
-
-class ZoomSpeed(IntEnum):
-    ZOOM_OUT_FAST = -200
-    ZOOM_OUT_SLOW = -100
-    ZOOM_STOPPED = 0
-    ZOOM_IN_SLOW = 100
-    ZOOM_IN_FAST = 200
-
-
-@dataclass
-class CameraSpeeds:
-    pan_speed: float = 0
-    """Speed in degree per second. Positive values mean clockwise,
-    negative values stand for counter clockwise moving direction from the
-    camera's point of view.
-    """
-
-    tilt_speed: float = 0
-    """Speed in degree per second. Positive values mean upwards,
-    negative values stand for downwards moving direction from the camera's
-    point of view.
-    """
-
-    zoom_speed: ZoomSpeed = ZoomSpeed.ZOOM_STOPPED
-    """Abstract speed unit, i.e. the actual speed depends on camera model.
-    Positive values mean camera should zoom in (larger values mean that camera
-    should zoom faster), negative values stand for zooming out.
-    """
-
-    def reset(self) -> None:
-        """Stop camera movements.
-        """
-        self.pan_speed = 0
-        self.tilt_speed = 0
-        self.zoom_speed = ZoomSpeed.ZOOM_STOPPED
 
 
 class TrackingStrategy(Protocol):
